+++ priv/templates/brando.install/config/brando.exs
@@ -18,7 +18,6 @@ config :eightyfour,
 
 config :brando,
   app_name: "<%= application_module %>",
-  auth_sleep_duration: 5000,
   endpoint: <%= application_module %>.Endpoint,
   otp_app: :<%= application_name %>,
   log_dir: Path.expand("./logs"),